Move ConstC and ConstV Properties
