+++ src/baseclasses/helper/archive_builder/biologic_archive.py
@@ -6,12 +6,11 @@ import numpy as np
 from nomad.units import ureg
 
 import baseclasses
-from baseclasses.chemical_energy.chronoamperometry import CAProperties
+from baseclasses.chemical_energy.chronoamperometry import CAProperties, ConstVProperties
 from baseclasses.chemical_energy.chronocoulometry import CCProperties
-from baseclasses.chemical_energy.chronopotentiometry import CPProperties
-from baseclasses.chemical_energy.constantpotential import (
+from baseclasses.chemical_energy.chronopotentiometry import (
     ConstCProperties,
-    ConstVProperties,
+    CPProperties,
 )
 from baseclasses.chemical_energy.cyclicvoltammetry import CVProperties
 from baseclasses.chemical_energy.electrochemical_impedance_spectroscopy import (
